backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
tortoise-orm[asyncpg]==0.21.7
aerich==0.7.2
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
slowapi==0.1.9
httpx==0.28.1
asyncpg==0.30.0
